Move executor Kafka utils to separate file in the same package
